requirements.txt
streamlit
openai
google-generativeai
pydub
pandas
weasyprint
json_repair